VAN-50 Ajinkya/bytesn (#43)
* [WIP] python: handle preprocessing of address send/transfer

- Add xfailed unit test for address send/transfer
- C++ component still crashes (assertion failure)

* ckp

* Added translations of intrinsics required for handling of dynamic byte array in memory

* storage ops moved from python to cpp

* Adds capability to handle dynamic storage and offset
Adds precise typed stored and loads

* Decomposed storage load stores int smaller modular functions.

* Removed cleanup functions

* Enabled bytes1 and bytes2 tests

+++ tests/e2e/feature/nested-mapping.sol
@@ -5,24 +5,24 @@ pragma solidity ^0.8.10;
 
 
 contract NestedMappingTestCase {
-    mapping(uint256 => mapping(uint256 => uint256)) public map;
+    mapping(uint256 => mapping(uint256 => uint32)) public map;
 
-    function mappingRead(uint256 x, uint256 y) external view returns (uint256) {
+    function mappingRead(uint256 x, uint256 y) external view returns (uint32) {
         return map[x][y];
     }
 
-    function mappingWrite(uint256 k1, uint256 k2, uint256 v) external {
+    function mappingWrite(uint256 k1, uint256 k2, uint32 v) external {
         map[k1][k2] = v;
     }
 }
 
-//CHECK: {{define i256 @fun_mappingRead_[0-9]+\(i256\* \%.+, .+, .+\)}}
-//CHECK: {{call i256\* @pyul_map_index\(i256\* \%.+, .+\)}}
-//CHECK: {{call i256\* @pyul_map_index\(i256\* \%.+, .+\)}}
-//CHECK: {{load i256, i256\*}}
+//CHECK: define i256 @fun_mappingRead_{{[0-9]+\(i256\* \%.+, .+, .+\)}}
+//CHECK: call i256* @pyul_map_index{{\(i256\* \%.+, .+\)}}
+//CHECK: call i256* @pyul_map_index{{\(i256\* \%.+, .+\)}}
+//CHECK: load i32, i32*
 
 
-//CHECK: {{define void @fun_mappingWrite_[0-9]+\(i256\* \%.+, .+, .+, .+\)}}
-//CHECK: {{call i256\* @pyul_map_index\(i256\* \%.+, .+\)}}
-//CHECK: {{call i256\* @pyul_map_index\(i256\* \%.+, .+\)}}
-//CHECK: {{store i256 .+, i256\*.*}}
+//CHECK: define void @fun_mappingWrite_{{[0-9]+\(i256\* \%.+, .+, .+, .+\)}}
+//CHECK: call i256* @pyul_map_index({{i256\* \%.+, .+}})
+//CHECK: call i256* @pyul_map_index({{i256\* \%.+, .+}})
+//CHECK: store i32 {{.+}}, i32*
